+++ DataAccess/Entities/GpsData.cs
@@ -7,9 +7,9 @@ namespace DataAccess.Entities
         public double Latitude { get; set; }
         public double Longitude { get; set; }
         public DateTime GpsTime { get; set; }
-        public double Speed { get; set; }
-        public float Angle { get; set; }
-        public double Altitude { get; set; }
-        public int Satellites { get; set; }
+        public short Speed { get; set; }
+        public short Angle { get; set; }
+        public short Altitude { get; set; }
+        public byte Satellites { get; set; }
     }
 }
